chore: movimentação de algumas variáveis para arquivos separados na tela PlanSelectionScreen
