--- frontend/src/screens/onboarding/planSelection/types/types.ts
@@ -1,1 +1,0 @@
-export type PlanType = "gradual" | "moderado" | "acelerado";
